Fixed test for /backend/internal/db/roster.go
