+++ backend/internal/db/roster.go
@@ -17,19 +17,3 @@ func (s session) CreateRoster(roster *models.Roster) (*models.Roster, error) {
 
 	return resultOrError(roster, result)
 }
-
-func (s session) GetUserByID(id uint) (*models.User, error) {
-	user := &models.User{}
-
-	err := s.Find(&user, "ID = ?", id)
-
-	return resultOrError(user, err)
-}
-
-func (s session) GetUsersByIDs(ids []uint) ([]*models.User, error) {
-	users := make([]*models.User, 0)
-
-	err := s.Where("ID IN (?)", ids).Find(&users)
-
-	return resultsOrError(users, err)
-}
