+++ backend/internal/db/roster_test.go
@@ -12,7 +12,8 @@ func (s *dbTestingSuite) TestCreateRoster() {
 
 	date, _ := time.Parse(layout, value)
 
-	captain = models.User{
+	// Create the Captain of the Roster
+	captain := models.User{
 		FirstName:    "John",
 		LastName:     "Smith",
 		Email:        "test@email.com",
@@ -24,15 +25,16 @@ func (s *dbTestingSuite) TestCreateRoster() {
 		DateOfBirth:  date,
 	}
 
-	dbCapt, err := s.session.CreateUser(captain)
+	dbCapt, err := s.session.CreateUser(&captain)
 	s.Nil(err)
 	s.NotNil(dbCapt)
 
-	capt, err := s.session.GetUserbyID(dbCapt.ID)
+	capt, err := s.session.GetUserByID(dbCapt.ID)
 	s.Nil(err)
 	s.NotNil(capt)
 
-	user1 = models.User{
+	// Create the list of Players
+	user1 := models.User{
 		FirstName:    "Jack",
 		LastName:     "Smith",
 		Email:        "test1@email.com",
@@ -44,7 +46,7 @@ func (s *dbTestingSuite) TestCreateRoster() {
 		DateOfBirth:  date,
 	}
 
-	user2 = models.User{
+	user2 := models.User{
 		FirstName:    "Jill",
 		LastName:     "Smith",
 		Email:        "test2@email.com",
@@ -56,34 +58,39 @@ func (s *dbTestingSuite) TestCreateRoster() {
 		DateOfBirth:  date,
 	}
 
-	dbUser1, err := s.session.CreateUser(user1)
+	dbUser1, err := s.session.CreateUser(&user1)
 	s.Nil(err)
 	s.NotNil(dbUser1)
 
-	dbUser2, err := s.session.CreateUser(user2)
+	dbUser2, err := s.session.CreateUser(&user2)
 	s.Nil(err)
 	s.NotNil(dbUser2)
 
-	playerIDs := [2]uint{}
+	playerIDs := []uint{dbUser1.ID, dbUser2.ID}
 
 	players, err := s.session.GetUsersByIDs(playerIDs)
 	s.Nil(err)
 	s.NotNil(players)
 
+	// Create the Roster
 	roster := models.Roster{
 		CaptainID: capt.ID,
 		Captain:   *capt,
-		PlayerIDs: players,
+		Players:   players,
 	}
 
-	r, err := s.session.CreateRoster(&roster)
+	myRoster, err := s.session.CreateRoster(&roster)
 	s.Nil(err)
-	s.notNil(r)
+	s.NotNil(myRoster)
 
-	s.Equal(uint(1), r.ID)
+	// Verify the number of players on the team besides the Captain
+	// equals 2
+	s.Len(myRoster.Players, 2)
 
+	s.Equal(uint(1), myRoster.ID)
+
+	// Test Getting Rosters
 	rosters, err := s.session.GetRosters()
 	s.Nil(err)
-
 	s.Len(rosters, 1)
 }
